DELETE - reparat - mutat useState din CityCard in Itinerary, mutat functie delete in Itinerary, chestiune decisiva functionalitate: fortare rerender useEffect useFetchData cu clicked din functia delete; redenumire index in >>ind<< la mapare data din CityCard
