+++ src/components/CityCard/CityCard.jsx
@@ -2,23 +2,29 @@ import { useContext, useState } from "react";
 import Button from "react-bootstrap/Button";
 import Card from "react-bootstrap/Card";
 import Modal from "react-bootstrap/Modal";
-import { itineraryMinus } from "../../global/itinerary/actions";
+// import { itineraryMinus } from "../../global/itinerary/actions";
 import { ItineraryContext } from "../../global/itinerary/context";
 import useFetchData from "../../hooks/useFetchData";
 import { DeleteButton } from "../CityCard/CityCard.style";
 import { ImgWrapper } from "../DestinationCard/DestinationCard.style";
 
-function CityCard({ country, city, index }) {
+function CityCard({
+  country,
+  city,
+  index,
+  handleDelete,
+  show,
+  setShow,
+  setShowId,
+  clicked,
+  setClicked,
+}) {
   console.log("country", country, "city", city, "index", index);
 
-  const { stateGlobalItinerary, dispatchItinerary } =
-    useContext(ItineraryContext);
+  const { stateGlobalItinerary } = useContext(ItineraryContext);
 
   const itineraryValueArray = stateGlobalItinerary.itineraryValue;
   console.log("itineraryValueArray", itineraryValueArray);
-
-  const [clicked, setClicked] = useState(true);
-  const [show, setShow] = useState(0);
 
   const url = `http://localhost:3001/${country}?city=${city}`;
   console.log("url", url);
@@ -26,13 +32,8 @@ function CityCard({ country, city, index }) {
   const { data, error, loading } = useFetchData(url, clicked, setClicked);
   console.log("data", "error", "loading", data, error, loading);
 
-  const handleDelete = (index) => {
-    dispatchItinerary(itineraryMinus(index));
-    setShow(false);
-  };
-
   const handleCloseShow = () => {
-    setShow(!show);
+    setShow(!show), setShowId(index);
   };
   return (
     <>
@@ -55,8 +56,8 @@ function CityCard({ country, city, index }) {
         </Modal.Footer>
       </Modal>
       {data &&
-        data?.map((card, index) => (
-          <Card key={index} className="tangerine-bold">
+        data?.map((card, ind) => (
+          <Card key={ind} className="tangerine-bold">
             <ImgWrapper loc="ImgWrapper">
               <Card.Img variant="top" src={card.image} alt="Image of city" />
             </ImgWrapper>
@@ -65,7 +66,7 @@ function CityCard({ country, city, index }) {
               <Card.Text>{card.description}</Card.Text>
               <DeleteButton
                 loc="DeleteButton"
-                onClick={() => handleCloseShow(index)}
+                onClick={() => handleCloseShow()}
               >
                 Remove from My Itinerary
               </DeleteButton>
