+++ src/components/Itinerary/Itinerary.jsx
@@ -1,4 +1,4 @@
-import { useContext } from "react";
+import { useContext, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { ItineraryContext } from "../../global/itinerary/context";
 import useLocalStorage from "../../hooks/useLocalStorage";
@@ -10,10 +10,12 @@ import {
   ButtonAccommodation,
   SectionItinerary,
 } from "./Itinerary.style";
+import { itineraryMinus } from "../../global/itinerary/actions";
 
 function Itinerary() {
   const navigate = useNavigate();
-  const { stateGlobalItinerary } = useContext(ItineraryContext);
+  const { stateGlobalItinerary, dispatchItinerary } =
+    useContext(ItineraryContext);
 
   const itineraryValueArray = stateGlobalItinerary.itineraryValue || [];
   console.log("itineraryValueArray", itineraryValueArray);
@@ -24,6 +26,10 @@ function Itinerary() {
 
   const { localData } = useLocalStorage("user");
   console.log("localData", localData);
+
+  const [clicked, setClicked] = useState(true);
+  const [show, setShow] = useState(false);
+  const [showId, setShowId] = useState(0);
 
   let accommodationArray = [];
 
@@ -73,12 +79,28 @@ function Itinerary() {
     });
   };
 
+  const handleDelete = (index) => {
+    dispatchItinerary(itineraryMinus(index));
+    setClicked(true);
+    setShow(false);
+  };
+
   return (
     <SectionItinerary loc="SectionItinerary">
       <ItineraryData loc="ItineraryData">
         {stateGlobalItinerary &&
           itineraryValueArray?.map((element, index) => (
-            <CityCard key={index} index={index} {...element} />
+            <CityCard
+              key={index}
+              index={index}
+              handleDelete={handleDelete}
+              show={show && showId === index}
+              setShow={setShow}
+              setShowId={setShowId}
+              clicked={clicked}
+              setClicked={setClicked}
+              {...element}
+            />
           ))}
         {stateGlobalItinerary &&
           itineraryLandmarkValueArray?.map((element, index) => (
